apps/backend/requirements.txt
fastapi==0.115.6
openai==1.68.2
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.33.0
